refactor: 레거시 MVC 제거 및 패키지 구조 변경
