+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/asis/ManualHandlerMapping.java
@@ -1,11 +1,9 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet.mvc.asis;
 
 import jakarta.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webmvc.org.springframework.web.servlet.mvc.HandlerMapping;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-import webmvc.org.springframework.web.servlet.mvc.asis.ForwardController;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -15,7 +13,6 @@ public class ManualHandlerMapping implements HandlerMapping {
     private static final Map<String, Controller> controllers = new HashMap<>();
 
     public void initialize() {
-        controllers.put("/", new ForwardController("/index.jsp"));
         log.info("Initialized Handler Mapping!");
         controllers.keySet()
                 .forEach(path -> log.info("Path : {}, Controller : {}", path, controllers.get(path).getClass()));
@@ -26,5 +23,9 @@ public class ManualHandlerMapping implements HandlerMapping {
         final String requestURI = request.getRequestURI();
         log.debug("Request Mapping Uri : {}", requestURI);
         return controllers.get(requestURI);
+    }
+
+    public void addController(final String url, final Controller controller) {
+        controllers.put(url, controller);
     }
 }
